+++ Addons/Schumix.CalendarAddon/Commands/CalendarCommand.cs
@@ -2,7 +2,7 @@
  * This file is part of Schumix.
  * 
  * Copyright (C) 2010-2013 Megax <http://megax.yeahunter.hu/>
- * Copyright (C) 2013 Schumix Team <http://schumix.eu/>
+ * Copyright (C) 2013-2014 Schumix Team <http://schumix.eu/>
  * 
  * Schumix is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -95,7 +95,13 @@ namespace Schumix.CalendarAddon.Commands
 
 		private string GetMessage(string args)
 		{
-			return _regex.IsMatch(args) ? _regex.Match(args).Groups["text"].ToString() : string.Empty;
+			if(_regex.IsMatch(args) && IsYear(args))
+				return _regex.Match(args).Groups["text"].ToString();
+
+			if(_hamregex.IsMatch(args) && !IsYear(args))
+				return _hamregex.Match(args).Groups["text"].ToString();
+
+			return string.Empty;
 		}
 
 		private bool IsYear(string args)