+++ Addons/Schumix.GameAddon/MaffiaGame/MaffiaGame.cs
@@ -2,7 +2,7 @@
  * This file is part of Schumix.
  * 
  * Copyright (C) 2010-2013 Megax <http://megax.yeahunter.hu/>
- * Copyright (C) 2013 Schumix Team <http://schumix.eu/>
+ * Copyright (C) 2013-2014 Schumix Team <http://schumix.eu/>
  * 
  * Schumix is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -391,7 +391,7 @@ namespace Schumix.GameAddon.MaffiaGames
 			}
 
 			if(_owner == OldName.ToLower())
-				_owner = NewName;
+				_owner = NewName.ToLower();
 
 			if(Started)
 			{