+++ Core/Schumix.Irc/Util/StatsQuery.cs
@@ -4,7 +4,7 @@
  * Copyright (C) 2002 Aaron Hunter <thresher@sharkbite.org>
  * Copyright (C) 2010-2012 Twl
  * Copyright (C) 2010-2013 Megax <http://megax.yeahunter.hu/>
- * Copyright (C) 2013 Schumix Team <http://schumix.eu/>
+ * Copyright (C) 2013-2014 Schumix Team <http://schumix.eu/>
  * 
  * Schumix is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by