+++ Tests/Schumix.Framework.Test/Options/OptionSetTest.cs
@@ -6,7 +6,7 @@
  *
  * Copyright (c) 2008 Novell, Inc. (http://www.novell.com)
  * Copyright (C) 2010-2013 Megax <http://megax.yeahunter.hu/>
- * Copyright (C) 2013 Schumix Team <http://schumix.eu/>
+ * Copyright (C) 2013-2014 Schumix Team <http://schumix.eu/>
  * 
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
@@ -199,7 +199,7 @@ namespace Schumix.Framework.Options.Test
 			Assert.AreEqual(libs[0], "/foo");
 			Assert.AreEqual(libs[1], null);
 
-			Utils.AssertException(typeof(OptionException), "Cannot bundle unregistered option '-V'.", p, v => {
+			Utils.AssertException(typeof(OptionException), "Cannot use unregistered option 'V' in bundle '-EVALUENOTSUP'.", p, v => {
 				v.Parse(_("-EVALUENOTSUP")); });
 		}
 
@@ -370,7 +370,7 @@ namespace Schumix.Framework.Options.Test
 				v.Parse(_("--f", "invalid")); });
 
 			// try to bundle with an option requiring a value
-			Utils.AssertException(typeof(OptionException), "Cannot bundle unregistered option '-z'.", p, v => {
+			Utils.AssertException(typeof(OptionException), "Cannot use unregistered option 'z' in bundle '-cz'.", p, v => {
 				v.Parse(_("-cz", "extra")); });
 
 			Utils.AssertException(typeof(ArgumentNullException), "Argument cannot be null.\nParameter name: action", p, v => {
@@ -862,6 +862,22 @@ namespace Schumix.Framework.Options.Test
 			Assert.AreEqual(formats["baz"].Count, 2);
 			Assert.AreEqual(formats["baz"][0], "e");
 			Assert.AreEqual(formats["baz"][1], "f");
+		}
+
+		[Test]
+		public void ReportInvalidDuplication ()
+		{
+			int verbosity = 0;
+			var p = new OptionSet () {
+				{ "v", v => verbosity = v != null ? verbosity + 1 : verbosity },
+			};
+			try {
+				p.Parse (new []{"-v-v-v"});
+				Assert.Fail ("Should not be reached.");
+			} catch (OptionException e) {
+				Assert.AreEqual (null, e.OptionName);
+				Assert.AreEqual ("Cannot use unregistered option '-' in bundle '-v-v-v'.", e.Message);
+			}
 		}
 	}
 }